requirements.txt
google-genai
matplotlib
python-dotenv
requests
Pillow
google-cloud-storage
google-cloud-bigquery
streamlit
pandas
db-dtypes
numpy<2.0
